requirements.txt
protobuf
numpy
tensorflow
tushare
bottle
pyyaml